Raspberry Pi Motion Capture Python Script and Grive Syncing
This is a remote security camera setup for a Raspberry Pi Computer and
Pi Camera Module.  The main motion capture program is written in python.
Grive and associated script will synchronize the motion capture images
with a google drive account to allow access to images and internet
connected device or computer using google drive app.